Se separan responsabilidades del módulo principal y se cubren los casos cuando alguno de los apellidos es nil
