+++ mix.exs
@@ -4,7 +4,7 @@ defmodule RfcFacil.MixProject do
   def project do
     [
       app: :rfc_facil,
-      version: "0.1.3",
+      version: "0.1.4",
       elixir: "~> 1.9",
       package: package(),
       start_permanent: Mix.env() == :prod,
